server.R;bugfix;Wow, here's a failure mode to be aware of going forward...
Behavior:
Web page goes gray a few seconds after rendering without any user input.

Console:
Warning: Error in py_get_attr_impl: AttributeError: 'module' object has no attribute 'dfmeta'
  84: <Anonymous>

Cause:
When I changed everything over to the py environment instead of the rvp reactivValues object, the behavior changed when referencing a non-yet-existing object: instead of R returning NULL and continuing there is a Python AttributError getting raised. It's easy enough to fix, just check for existence before referencing it, but it wasn't obvious from the console message what the problem was and where it was happening (in the code that checks for uniqueness/validity of custom rules as it happens). This is why I'm putting in an especially verbose commit message-- to remind future me and future collaborators that an AttributeError could be caused by py being referenced in a reactive context before the needed object exists, and can be fixed by manually checking for it first.